Move data frame utils to their own C file
