+++ src/utils.c
@@ -165,42 +165,6 @@ SEXP df_map(SEXP df, SEXP (*fn)(SEXP)) {
   return out;
 }
 
-bool is_compact_rownames(SEXP x) {
-  return Rf_length(x) == 2 && INTEGER(x)[0] == NA_INTEGER;
-}
-R_len_t compact_rownames_length(SEXP x) {
-  return abs(INTEGER(x)[1]);
-}
-
-static void init_compact_rownames(SEXP x, R_len_t n);
-static SEXP new_compact_rownames(R_len_t n);
-
-void init_data_frame(SEXP x, R_len_t n) {
-  Rf_setAttrib(x, R_ClassSymbol, classes_data_frame);
-  init_compact_rownames(x, n);
-}
-void init_tibble(SEXP x, R_len_t n) {
-  Rf_setAttrib(x, R_ClassSymbol, classes_tibble);
-  init_compact_rownames(x, n);
-}
-
-static void init_compact_rownames(SEXP x, R_len_t n) {
-  SEXP rn = PROTECT(new_compact_rownames(n));
-  Rf_setAttrib(x, R_RowNamesSymbol, rn);
-  UNPROTECT(1);
-}
-static SEXP new_compact_rownames(R_len_t n) {
-  if (n <= 0) {
-    return vctrs_shared_empty_int;
-  }
-
-  SEXP out = Rf_allocVector(INTSXP, 2);
-  int* out_data = INTEGER(out);
-  out_data[0] = NA_INTEGER;
-  out_data[1] = -n;
-  return out;
-}
-
 inline void never_reached(const char* fn) {
   Rf_error("Internal error in `%s()`: Never reached", fn);
 }
